Added suggestions and remove planner for ruby
