+++ internal/pkgsuggest/pkgsuggest.go
@@ -8,6 +8,7 @@ import (
 	"go.jetpack.io/devbox/internal/pkgsuggest/suggestors/java"
 	"go.jetpack.io/devbox/internal/pkgsuggest/suggestors/javascript"
 	"go.jetpack.io/devbox/internal/pkgsuggest/suggestors/nginx"
+	"go.jetpack.io/devbox/internal/pkgsuggest/suggestors/ruby"
 	"go.jetpack.io/devbox/internal/pkgsuggest/suggestors/rust"
 	"go.jetpack.io/devbox/internal/pkgsuggest/suggestors/zig"
 )
@@ -19,6 +20,7 @@ var SUGGESTORS = []suggestors.Suggestor{
 	&java.Suggestor{},
 	&javascript.Suggestor{},
 	&nginx.Suggestor{},
+	&ruby.Suggestor{},
 	&rust.Suggestor{},
 	&zig.Suggestor{},
 }
