+++ internal/planner/planner.go
@@ -7,7 +7,6 @@ import (
 	"github.com/samber/lo"
 	"go.jetpack.io/devbox/internal/planner/languages/php"
 	"go.jetpack.io/devbox/internal/planner/languages/python"
-	"go.jetpack.io/devbox/internal/planner/languages/ruby"
 	"go.jetpack.io/devbox/internal/planner/plansdk"
 )
 
@@ -15,7 +14,6 @@ var PLANNERS = []plansdk.Planner{
 	&php.V2Planner{},
 	&python.PoetryPlanner{},
 	&python.PIPPlanner{},
-	&ruby.Planner{},
 }
 
 // Return a merged shell plan from shell planners if user defined packages
